Safari-609.1.11.1/LayoutTests/media/modern-media-controls/rewind-button/rewind-button.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>RewindButton</code> class.");

const rewindButton = new RewindButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("rewindButton.element.localName", "button");
shouldBeTrue('rewindButton.element.classList.contains("rewind")');
shouldBe("rewindButton.iconName", "Icons.Rewind");

shouldBecomeEqual('rewindButton.image.element.style.webkitMaskImage.includes("macOS/Rewind.svg")', "true", finishJSTest);

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
