Safari-609.1.11.1/LayoutTests/media/modern-media-controls/rewind-button/rewind-button-expected.txt

Testing the RewindButton class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rewindButton.element.localName is "button"
PASS rewindButton.element.classList.contains("rewind") is true
PASS rewindButton.iconName is Icons.Rewind
PASS rewindButton.image.element.style.webkitMaskImage.includes("macOS/Rewind.svg") became true
PASS successfullyParsed is true

TEST COMPLETE
